Reapply spilt file again
