+++ src/components/discover-sheet/DiscoverSheet.ios.js
@@ -1,8 +1,6 @@
-import BottomSheet, { BottomSheetScrollView } from '@gorhom/bottom-sheet';
 import { useIsFocused } from '@react-navigation/native';
 import React, {
   forwardRef,
-  useCallback,
   useEffect,
   useImperativeHandle,
   useMemo,
@@ -10,22 +8,13 @@ import React, {
   useState,
 } from 'react';
 import { findNodeHandle, NativeModules } from 'react-native';
-import {
-  runOnJS,
-  useAnimatedReaction,
-  useAnimatedScrollHandler,
-  useSharedValue,
-} from 'react-native-reanimated';
 // eslint-disable-next-line import/no-unresolved
 import SlackBottomSheet from 'react-native-slack-bottom-sheet';
-import styled from 'styled-components';
 import { SlackSheet } from '../sheet';
 import DiscoverSheetContent from './DiscoverSheetContent';
 import DiscoverSheetContext from './DiscoverSheetContext';
 import DiscoverSheetHeader from './DiscoverSheetHeader';
-import CustomBackground from './androidCustomComponents/customBackground';
-import CustomHandle from './androidCustomComponents/customHandle';
-import { deviceUtils, safeAreaInsetValues } from '@rainbow-me/utils';
+import { safeAreaInsetValues } from '@rainbow-me/utils';
 
 const renderHeader = yPosition => <DiscoverSheetHeader yPosition={yPosition} />;
 
@@ -34,104 +23,7 @@ function useAreHeaderButtonVisible() {
   return [{ isSearchModeEnabled, setIsSearchModeEnabled }, isSearchModeEnabled];
 }
 
-const snapPoints = ['25%', '90%'];
-
-const AndroidWrapper = styled.View.attrs({
-  pointerEvents: 'box-none',
-})`
-  width: 100%;
-  height: ${deviceUtils.dimensions.height};
-  position: absolute;
-`;
-
-const DiscoverSheetAndroid = (_, forwardedRef) => {
-  const [headerButtonsHandlers, deps] = useAreHeaderButtonVisible();
-
-  const listeners = useRef([]);
-  const bottomSheetModalRef = useRef(null);
-  const onFabSearch = useRef(null);
-  const sheet = useRef();
-
-  const value = useMemo(
-    () => ({
-      addOnCrossMagicBorderListener(listener) {
-        listeners.current.push(listener);
-        return () =>
-          listeners.current.splice(listeners.current.indexOf(listener), 1);
-      },
-      jumpToLong() {
-        bottomSheetModalRef.current.expand();
-      },
-      jumpToShort() {
-        sheet.current.scrollTo({ animated: false, x: 0, y: 0 });
-        bottomSheetModalRef.current.collapse();
-      },
-      onFabSearch,
-      ...headerButtonsHandlers,
-    }),
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [deps]
-  );
-
-  useImperativeHandle(forwardedRef, () => value);
-
-  const yPosition = useSharedValue(0);
-  const scrollHandler = useAnimatedScrollHandler(event => {
-    yPosition.value = event.contentOffset.y;
-  });
-
-  const notifyListeners = useCallback(
-    value => listeners.current.forEach(listener => listener(value)),
-    []
-  );
-
-  const sheetPosition = useSharedValue(0);
-
-  const hasCrossedTheMagicBorder = useSharedValue(false);
-  useAnimatedReaction(
-    () => sheetPosition.value > 150,
-    hasJustCrossedTheMagicBorder => {
-      if (hasCrossedTheMagicBorder.value !== hasJustCrossedTheMagicBorder) {
-        runOnJS(notifyListeners)(hasJustCrossedTheMagicBorder);
-        hasCrossedTheMagicBorder.value = hasJustCrossedTheMagicBorder;
-      }
-    },
-    [notifyListeners]
-  );
-
-  return (
-    <AndroidWrapper
-      style={{
-        bottom: -24,
-      }}
-    >
-      <DiscoverSheetContext.Provider value={value}>
-        <BottomSheet
-          activeOffsetY={[-3, 3]}
-          animatedPosition={sheetPosition}
-          animationDuration={300}
-          backgroundComponent={CustomBackground}
-          failOffsetX={[-10, 10]}
-          handleComponent={CustomHandle}
-          index={1}
-          ref={bottomSheetModalRef}
-          snapPoints={snapPoints}
-          style={{
-            borderRadius: 20,
-            overflow: 'hidden',
-          }}
-        >
-          <DiscoverSheetHeader yPosition={yPosition} />
-          <BottomSheetScrollView onScroll={scrollHandler} ref={sheet}>
-            <DiscoverSheetContent />
-          </BottomSheetScrollView>
-        </BottomSheet>
-      </DiscoverSheetContext.Provider>
-    </AndroidWrapper>
-  );
-};
-
-function DiscoverSheetIOS(_, forwardedRef) {
+function DiscoverSheet(_, forwardedRef) {
   const isFocused = useIsFocused();
   const sheet = useRef();
   const ref = useRef();
@@ -216,4 +108,4 @@ function DiscoverSheetIOS(_, forwardedRef) {
   );
 }
 
-export default forwardRef(ios ? DiscoverSheetIOS : DiscoverSheetAndroid);
+export default forwardRef(DiscoverSheet);
